+++ tests/quote_unquote.rs
@@ -1,7 +1,7 @@
 use call2_for_syn::{call2_allow_incomplete, call2_strict};
-use proc_macro2::Span;
-use quote::quote;
-use syn::{parse2, Ident, Lit, LitStr, Result, Token};
+use proc_macro2::{Span, TokenStream};
+use quote::{quote, ToTokens};
+use syn::{parse::ParseStream, parse2, Attribute, Ident, Lit, LitStr, Result, Token};
 use unquote::unquote;
 
 //FIXME: These tests should also evaluate failures, but `call2` currently panics if not all input was parsed.
@@ -118,6 +118,40 @@ fn span_range() -> Result<()> {
 		Result::Ok(span)
 	})
 	.unwrap()?;
+
+	Ok(())
+}
+
+#[derive(Debug)]
+struct Attributes(Vec<Attribute>);
+impl Attributes {
+	fn parse_outer(input: ParseStream) -> Result<Self> {
+		Attribute::parse_outer(input).map(Self)
+	}
+}
+impl ToTokens for Attributes {
+	fn to_tokens(&self, tokens: &mut TokenStream) {
+		for attr in self.0.iter() {
+			attr.to_tokens(tokens)
+		}
+	}
+}
+
+#[test]
+fn r#do() -> Result<()> {
+	let tokens = quote! {
+		#[some_attribute]
+		#[another_attribute]
+	};
+
+	let attrs = call2_strict(tokens, |input| {
+		let attr;
+		unquote!(input, #do Attributes::parse_outer => attr);
+		Result::Ok(attr)
+	})
+	.unwrap()?;
+
+	assert_eq!(attrs.0.len(), 2);
 
 	Ok(())
 }