+++ src/lib.rs
@@ -1,4 +1,4 @@
-#![doc(html_root_url = "https://docs.rs/unquote/0.0.5")]
+#![doc(html_root_url = "https://docs.rs/unquote/0.0.6")]
 #![warn(clippy::pedantic)]
 
 #[cfg(doctest)]