+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 TODO: Date
 
 * Use `#let ident` to declare `ident`.
+* Use `#do parser_function => binding` to parse `binding` via `parser_function`.
 
 ## 0.0.5